refactor(rag): タイムアウト定数を共通モジュールに集約
PR #508 レビュー指摘対応: lib/rag/agent-timeouts.ts に
SERVER_MAX_DURATION_MS, FALLBACK_MARGIN_MS, AGENT_TIMEOUT_MS,
CLIENT_TIMEOUT_MS を集約し、schemas.ts と hooks から参照。
