+++ app/api/rag/agent-search/batch-handler.ts
@@ -9,7 +9,7 @@ import { SpanStatusCode, Span } from '@opentelemetry/api';
 import type { Session } from 'next-auth';
 
 import type { RateLimitInfo, ValidatedRequest, ModeContext } from './schemas';
-import { AGENT_TIMEOUT_MS } from './schemas';
+import { AGENT_TIMEOUT_MS } from '@/lib/rag/agent-timeouts';
 import {
   attachRateLimitHeaders,
   unwrapToolOutput,
